refactor losses and create plugin architecture
