+++ openpifpaf/__init__.py
@@ -17,6 +17,7 @@ from . import plugin
 
 from .datasets import DATAMODULES
 from .decoder import DECODERS
-from .network.factory import BASE_FACTORIES, BASE_TYPES, CHECKPOINT_URLS, HEAD_FACTORIES, HEAD_TYPES
+from .network.factory import BASE_FACTORIES, BASE_TYPES, CHECKPOINT_URLS, HEADS
+from .network.losses.factory import LOSSES, LOSS_COMPONENTS
 from .network.nets import MODEL_MIGRATION
 from .show.annotation_painter import PAINTERS
