+++ openpifpaf/network/losses/__init__.py
@@ -1,0 +1,1 @@
+from .factory import cli, configure, factory, factory_from_args
